Rework VectorUI class task 1 level b

+++ HomeworkStage14/src/by/itstep/oris/workspace/leveld/task01/view/VectorUI.java
@@ -3,9 +3,9 @@ package by.itstep.oris.workspace.leveld.task01.view;
 import java.util.Arrays;
 
 public class VectorUI {
-	public static void getResult(int sum, int... vector) {
+	public static void getResult(double sum, double... vector) {
 		System.out.println(String.format
-				("Sum of elements between min and max elements in %s is: %d",
+				("Sum of elements between min and max elements in %s is: %.2f",
 						 Arrays.toString(vector), sum));
 	}
 }
